+++ lib/widget/bmi_results_widget.dart
@@ -1,25 +1,21 @@
+import 'package:auto_size_text/auto_size_text.dart';
 import 'package:bmiapp/components/app_color.dart';
 import 'package:bmiapp/components/app_text.dart';
 import 'package:bmiapp/components/app_textstyle.dart';
 import 'package:flutter/material.dart';
 
-class BmiResults extends StatefulWidget {
+class BmiResults extends StatelessWidget {
   const BmiResults({super.key});
 
-  @override
-  State<BmiResults> createState() => _BmiResultsState();
-}
-
-class _BmiResultsState extends State<BmiResults> {
-  
   @override
   Widget build(BuildContext context) {
     final screenSize = MediaQuery.of(context).size;
     final screenHeight = screenSize.height;
     final screenWidth = screenSize.width;
-    final args = ModalRoute.of(context)!.settings.arguments as Map<String, int>;
-    final height = args['height']!;
-    final weight = args['weight']!;
+    final args =
+        ModalRoute.of(context)?.settings.arguments as Map<String, int>?;
+    final height = args?['height'] ?? 0;
+    final weight = args?['weight'] ?? 0;
     final bmi = weight / ((height / 100) * (height / 100));
     String status;
     if (bmi < 18.5) {
@@ -32,52 +28,49 @@ class _BmiResultsState extends State<BmiResults> {
       status = 'Obese';
     }
     return Container(
-              width: screenWidth - 60,
-              height: (413 / 852) * screenHeight,
-              decoration: BoxDecoration(
-                color: AppColor.white,
-                borderRadius: BorderRadius.circular(12),
+      width: screenWidth - 60,
+      height: (413 / 852) * screenHeight,
+      decoration: BoxDecoration(
+        color: AppColor.white,
+        borderRadius: BorderRadius.circular(12),
+      ),
+      child: Column(
+        children: [
+          SizedBox(height: 56),
+          AppText(
+            text: 'BMI Results',
+            style: AppTextstyle.tsRegularMidnightBlue32,
+          ),
+          Row(
+            crossAxisAlignment: CrossAxisAlignment.baseline,
+            textBaseline: TextBaseline.alphabetic,
+            children: [
+              SizedBox(width: 50),
+              Flexible(
+                child: AutoSizeText(
+                  bmi.toStringAsFixed(0),
+                  style: AppTextstyle.tsBoldVioletBlue140.copyWith(height: 1),
+                  maxLines: 1,
+                ),
               ),
-              child: Column(
-                children: [
-                  SizedBox(height: 56),
-                  AppText(
-                    text: 'BMI Results',
-                    style: AppTextstyle.tsRegularMidnightBlue32,
-                  ),
-                  Row(
-                    crossAxisAlignment: CrossAxisAlignment.baseline,
-                    textBaseline: TextBaseline.alphabetic,
-                    children: [
-                      SizedBox(width: 50),
-                      AppText(
-                        text: bmi.toStringAsFixed(0),
-                        style: AppTextstyle.tsBoldVioletBlue140.copyWith(
-                          height: 1,
-                        ),
-                      ),
-                      AppText(
-                        text:
-                            '.${bmi.toStringAsFixed(5).split('.')[1].substring(0, 2)}',
-                        style: AppTextstyle.tsMediumVioletBlue42.copyWith(
-                          height: 1,
-                        ),
-                      ),
-                    ],
-                  ),
-                  SizedBox(height: 2.46),
-                  AppText(
-                    text: status,
-                    style: AppTextstyle.tsMediumMidnightBlue24,
-                  ),
-                  SizedBox(height: 18.54),
-                  AppText(
-                    text:
-                        'Underweight: BMI less than 18.5\nNormal weight: BMI 18.5 to 24.9\nOverweight: BMI 25 to 29.9\nObesity: 30 to 40',
-                    style: AppTextstyle.tsMediumMidnightBlue13,
-                  ),
-                ],
+              AutoSizeText(
+                '.${bmi.toStringAsFixed(5).split('.')[1].substring(0, 2)}',
+                style: AppTextstyle.tsMediumVioletBlue42.copyWith(height: 1),
+                maxLines: 1,
               ),
-            );
+              SizedBox(width: 30),
+            ],
+          ),
+          SizedBox(height: 2.46),
+          AppText(text: status, style: AppTextstyle.tsMediumMidnightBlue24),
+          SizedBox(height: 18.54),
+          AppText(
+            text:
+                'Underweight: BMI less than 18.5\nNormal weight: BMI 18.5 to 24.9\nOverweight: BMI 25 to 29.9\nObesity: 30 to 40',
+            style: AppTextstyle.tsMediumMidnightBlue13,
+          ),
+        ],
+      ),
+    );
   }
 }
