+++ pubspec.yaml
@@ -32,12 +32,10 @@ dependencies:
     sdk: flutter
   flutter_svg: ^2.2.0
 
-
-
-
   # The following adds the Cupertino Icons font to your application.
   # Use with the CupertinoIcons class for iOS style icons.
   cupertino_icons: ^1.0.8
+  auto_size_text: ^3.0.0
 
 dev_dependencies:
   flutter_test:
